tui/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hookwise-tui"
version = "1.0.0"
description = "Rich terminal UI for hookwise — Claude Code hooks framework"
requires-python = ">=3.10"
dependencies = [
    "textual>=1.0.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "anthropic>=0.40.0",
]

[project.scripts]
hookwise-tui = "hookwise_tui.__main__:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
